Refactor: 성능 개선
