+++ 이인복/BOJ/java/com/Week03/LV3_81303/Q81303.java
@@ -1,4 +1,4 @@
-package com.Week03.LV2_81303;
+package com.Week03.LV3_81303;
 
 import java.util.Arrays;
 import java.util.Stack;
